Split dungeon editor controls content model
